requirements.txt
mcp
playwright
imagehash
Pillow
